@@ -17,15 +17,17 @@
     </properties>
 
     <dependencies>
+
+
         <dependency>
-            <groupId>org.example</groupId>
-            <artifactId>utils</artifactId>
+            <groupId>com.example</groupId>
+            <artifactId>database</artifactId>
             <version>1.0-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
-            <groupId>com.example</groupId>
-            <artifactId>database</artifactId>
+            <groupId>org.example</groupId>
+            <artifactId>utils</artifactId>
             <version>1.0-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
@@ -43,8 +45,8 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <source>15</source>
-                    <target>15</target>
+                    <source>16</source>
+                    <target>16</target>
                 </configuration>
             </plugin>
         </plugins>
